Check for Lato and fall back to ggplot2 default if not available (#12)
* Check for Lato and fall back to ggplot2 default if not available

* Mock available fonts so test will work on systems without Lato

The GHA machines don't necessarily have Lato

* Skip visual tests if systemfonts is not available

It's not currently available for R 3.2 and 3.3, and only used by vdiffr, and
we're not even running the visual tests on these versions

* Add helper to deal with vdiffr availability issues

R CMD check fails if vdiffr isn't installed, even if we're not running the
visual tests (e.g. on github actions with R 3.2, where vdiffr is not currently
available). This change raises an error only if we're expecting to run visual
tests and vdiffr isn't available.

* Disable CRAN check for suggested packages on old versions of R

* Remove vdiffr helper per Nicole's feedback